Update Paging and styles

+++ src/BuildingBlocks/Examination.Dtos/SeedWork/PagedList.cs
@@ -12,7 +12,7 @@ namespace Examination.Dtos.SeedWork
         public List<T> Items { set; get; }
 
         public PagedList() { }
-        public PagedList(List<T> items, int count, int pageNumber, int pageSize)
+        public PagedList(List<T> items, long count, int pageNumber, int pageSize)
         {
             MetaData = new MetaData
             {
